add: HQ consumer

+++ cmd/get.go
@@ -28,7 +28,7 @@ func getCMDsFlags(getCmd *cobra.Command) {
 	getCmd.PersistentFlags().String("job", "", "Job name to use, will determine the path for the persistent queue, seencheck database, and WARC files.")
 	getCmd.PersistentFlags().IntP("workers", "w", 1, "Number of concurrent workers to run.")
 	getCmd.PersistentFlags().Int("max-concurrent-assets", 8, "Max number of concurrent assets to fetch PER worker. E.g. if you have 100 workers and this setting at 8, Zeno could do up to 800 concurrent requests at any time.")
-	getCmd.PersistentFlags().Uint8("max-hops", 0, "Maximum number of hops to execute.")
+	getCmd.PersistentFlags().Int("max-hops", 0, "Maximum number of hops to execute.")
 	getCmd.PersistentFlags().String("cookies", "", "File containing cookies that will be used for requests.")
 	getCmd.PersistentFlags().Bool("keep-cookies", false, "Keep a global cookie jar")
 	getCmd.PersistentFlags().Bool("headless", false, "Use headless browsers instead of standard GET requests.")
@@ -95,7 +95,7 @@ func getCMDsFlags(getCmd *cobra.Command) {
 	// Aliases shouldn't be used as proper flags nor declared in the config struct
 	// Aliases should be marked as deprecated to inform the user base
 	// Aliases values should be copied to the proper flag in the config/config.go:handleFlagsAliases() function
-	getCmd.PersistentFlags().Uint8("hops", 0, "Maximum number of hops to execute.")
+	getCmd.PersistentFlags().Int("hops", 0, "Maximum number of hops to execute.")
 	getCmd.PersistentFlags().MarkDeprecated("hops", "use --max-hops instead")
 	getCmd.PersistentFlags().MarkHidden("hops")
